+++ tests/test_similarity_config.rs
@@ -6,7 +6,7 @@ use mass_spectrometry::prelude::{
 
 #[test]
 fn hungarian_cosine_rejects_negative_tolerance() {
-    let result = HungarianCosine::new(1.0_f32, 1.0_f32, -0.1_f32);
+    let result = HungarianCosine::new(1.0, 1.0, -0.1);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NegativeTolerance)
@@ -15,7 +15,7 @@ fn hungarian_cosine_rejects_negative_tolerance() {
 
 #[test]
 fn modified_hungarian_cosine_rejects_negative_tolerance() {
-    let result = ModifiedHungarianCosine::new(1.0_f32, 1.0_f32, -0.1_f32);
+    let result = ModifiedHungarianCosine::new(1.0, 1.0, -0.1);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NegativeTolerance)
@@ -24,7 +24,7 @@ fn modified_hungarian_cosine_rejects_negative_tolerance() {
 
 #[test]
 fn entropy_similarity_rejects_negative_tolerance() {
-    let result = LinearEntropy::weighted(-0.01_f32);
+    let result = LinearEntropy::weighted(-0.01);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NegativeTolerance)
@@ -33,7 +33,7 @@ fn entropy_similarity_rejects_negative_tolerance() {
 
 #[test]
 fn hungarian_cosine_rejects_nan_mz_power() {
-    let result = HungarianCosine::new(f32::NAN, 1.0_f32, 0.1_f32);
+    let result = HungarianCosine::new(f64::NAN, 1.0, 0.1);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NonFiniteParameter("mz_power"))
@@ -42,7 +42,7 @@ fn hungarian_cosine_rejects_nan_mz_power() {
 
 #[test]
 fn modified_hungarian_cosine_rejects_infinite_intensity_power() {
-    let result = ModifiedHungarianCosine::new(1.0_f32, f32::INFINITY, 0.1_f32);
+    let result = ModifiedHungarianCosine::new(1.0, f64::INFINITY, 0.1);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NonFiniteParameter("intensity_power"))
@@ -51,7 +51,7 @@ fn modified_hungarian_cosine_rejects_infinite_intensity_power() {
 
 #[test]
 fn entropy_similarity_rejects_nan_tolerance() {
-    let result = LinearEntropy::unweighted(f32::NAN);
+    let result = LinearEntropy::unweighted(f64::NAN);
     assert!(matches!(
         result,
         Err(SimilarityConfigError::NonFiniteParameter("mz_tolerance"))
